% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovm.R
\name{autocovm}
\alias{autocovm}
\title{Lag-k autocovariance matrix for multivariate time series}
\usage{
autocovm(Y, k)
}
\arguments{
\item{Y}{A multivariate time series.}

\item{k}{The lag k.}
}
\value{
Autocovariace matrix at lag k.
}
\description{
This is an auxillary function to compute the autocovariance matrix for
multivariate time series at lag k.
}
\details{
Compute the autocovariance matrix of mutlivariate time series \eqn{Y} at
lag \eqn{k}.
}
\author{
Meng Cao \email{meng.cao@colostate.edu} and
   Wen Zhou \email{rickzhouwen@gmail.com}
}

