% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLX.R
\name{CLX}
\alias{CLX}
\title{CLX-test for two sample means}
\usage{
CLX(X, Y, alpha, DNAME)
}
\arguments{
\item{X}{The n x p data matrix from the sample 1}

\item{Y}{The n x p data matrix from the sample 2.}

\item{alpha}{The prescribed level of significance}

\item{DNAME}{Default input.}
}
\value{
Value of testing statistic, p-value, alternative hypothesis, and the name of
   testing procedure.
}
\description{
Testing the equality of two high dimensional mean vectors using the testing procedure
by Cai, Liu and Xia (2014).
}
\details{
Implementing testing procedure proposed by Cai, Liu, and Xia (2014) to test
   the equality of two sample high dimensional mean vectors under the assumption of
   sparsity of signals.
}
\author{
Tong He
}
\references{
T. Cai, W. Liu, and Y. Xia (2014). Two-sample test of high dimensional means under
   dependence. J. R. Statist. Soc. B. 76:349--372.
}

