% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CQ2.R
\name{CQ2}
\alias{CQ2}
\title{CQ-test for two sample means}
\usage{
CQ2(X, Y, DNAME)
}
\arguments{
\item{X}{The n x p data matrix from the sample 1}

\item{Y}{The n x p data matrix from the sample 2.}

\item{DNAME}{Default input.}
}
\value{
Value of testing statistic, alternative hypothesis, and the name of testing
   procedure.
}
\description{
Testing the equality of two high dimensional mean vectors using the testing procedure
by Chen and Qin (2010)
}
\details{
Implementing testing procedure proposed by Chen and Qin (2010) to test the
   equality of two sample high dimensional mean vectors under the assumption of sparsity
   of signals.
}
\author{
Tong He
}
\references{
S. Chen and Y. Qin (2010). A two-sample test for high-dimensional data with
   applications to gene-set testing. Ann. Statist. 38, 808-835
}

