% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCtest.R
\name{equalCovs}
\alias{equalCovs}
\title{LC-test for equality of high dimensional covariances}
\usage{
equalCovs(X, Y, alpha, DNAME)
}
\arguments{
\item{X}{The n x p data matrix from the sample 1}

\item{Y}{The n x p data matrix from the sample 2.}

\item{alpha}{The prescribed level of significance}

\item{DNAME}{Default input.}
}
\value{
Value of testing statistic, p-value, alternative hypothesis, and the name of testing procedure.
}
\description{
Testing the equality of two high dimensional covariance matrices using the testing procedure by Li and Chen (2012).
}
\details{
Implementing testing procedure proposed by Li and Chen (2012) to test the equality of two sample high dimensional covariance matrices.
}
\author{
Tong He
}
\references{
J. Li and S. Chen (2012). Two sample tests for high-dimensional covariance matrices. Ann. Statist. 40, 908--940
}

