% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDtest_data.R
\docType{data}
\name{GO26}
\alias{GO26}
\title{GO26}
\format{A list of two data objects \code{GO26$X} and \code{GO26$Y}.}
\usage{
data(GO26)
}
\description{
A list of two sample matrices sliced from GO:0016032, it is about viral reproduction
}
\keyword{datasets}

