% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDtest_data.R
\docType{data}
\name{GO54}
\alias{GO54}
\title{GO54}
\format{A list of two data objects \code{GO54$X} and \code{GO54$Y}.}
\usage{
data(GO54)
}
\description{
A list of two sample matrices sliced from GO:0034080,
it is about CenH3-containing nucleosome assembly at centromere.
}
\keyword{datasets}

