% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opbw.R
\name{opbw}
\alias{opbw}
\title{Optimal bandwidth to estimate long-run covariance}
\usage{
opbw(X)
}
\arguments{
\item{X}{\eqn{p x n} data matrix, with \eqn{p} time series of length \eqn{n}}
}
\value{
An optimal bandwidth.
}
\description{
This is an auxillary function to estimate the bandwidth that used to estimate the long run covariance for testing multivariate white nosies.
}
\author{
Meng Cao, Wen Zhou
}
\references{
J. Chang, Q. Yao, and W. Zhou (2016) Testing for high-dimensional white noise using maximum cross correlations. \emph{Biometrika}, to appear.
}

