% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{segment}
\usage{
segment(Y, mean_y, k, n, p)
}
\arguments{
\item{Y}{\eqn{p x n} data matrix, with \eqn{p} time series of length \eqn{n}}

\item{mean_y}{mean vector}

\item{k}{lag}

\item{n}{each time series has length \eqn{n}}

\item{p}{\eqn{p} time series}
}
\value{
b An optimal bandwidth.
}
\description{
segment
}
\author{
Meng Cao
}
\references{
J. Chang, Q. Yao, and W. Zhou (2016) Testing for high-dimensional white noise using maximum cross correlations. \emph{Biometrika}, to appear.
}

