% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wntest.R
\name{wntest}
\alias{wntest}
\title{Testing for high-dimensional white noise}
\usage{
wntest(Y, M, k_max = 10, kk, type = 1, alpha = 0.05, k0 = 10,
  delta = 1.5, opt = 1, lambda = 0.01, lambda_search = seq(1e-04, 0.01,
  length.out = 50), fold = 5, S1 = NULL, cv_opt = NULL)
}
\arguments{
\item{Y}{A p by n data matrix of p time series of length n.}

\item{M}{Number of bootstrap replicates, ex. 2000.}

\item{k_max}{The largest number of lags to be tested for white noise (default is 10).}

\item{kk}{A vector of lags to be tested (ex. kk = seq(2,10, by = 2)), scalar is allowed.}

\item{type}{Type of test to perform:
1: wntest, Testing for high-dimensional white noise using maximum cross-correlations. Biometrika, 104(1), pp.111-127.
2: test_LM, parameter needed: Y, k,
3: test_pre, parameter needed: Y, k_max, kk
4: test_TB, parameter needed: Y}

\item{alpha}{Level of significance (default is 0.05).}

\item{k0}{A parameter in time series PCA for pre-transformation (default is 10).}

\item{delta}{The thresholding parameter in time series PCA for pre-transformation (default is 1.5).}

\item{opt}{Options for transfermation:
1: perform transformation using fastclime to do the precision matrix estimation.
2: perform transformation using with sample covariance matrix.
3: perform transformation using clime with cross validation to do the precision matrix estimation.
4: perform transformation using fastclime with cross validation to do the precision matrix estimation.
else do not do transformation}

\item{lambda}{The tuning parameter used in package fastclime.  The default value is 0.1.}

\item{lambda_search}{The tuning parameters search for fastclime. Default seq(1e-4, 1e-2, length.out = 50).}

\item{fold}{Nnumber of folds used in corss validation. Default 5.}

\item{S1}{True contempaneous covariance matrix of the data if known in advance.}

\item{cv_opt}{Cross validation result to report (default value 1: minimun error).}
}
\value{
\item{res}{White noise test result (0) or not (1).}
\item{p_value}{P_value for the test result.}
\item{M1}{Square root of the estimated inverse covariance matrix if transformation has been applied.}
}
\description{
Methods to test high-dimensional white noise.
}
\examples{
library(expm)
p = 15
n = 300
S1 = diag(1, p, p)
for(ii in c(1:p)){
for(jj in c(1:p)){
S1[ii, jj] = 0.995^(abs(ii-jj))
}
}
S11 = sqrtm(S1)
X = S11 \%*\% matrix(rt(n*p, df = 8), ncol = n)
k_max = 10
kk = seq(2, k_max, 2)
M = 2000
k0 = 10
delta = 1.5
alpha = 0.05
wntest(X, M, k_max, kk, type = 1, opt = 0)
wntest(X, M, k_max, kk, type = 1, opt = 4, cv_opt = 1)

}
\references{
Chang, J., Yao, Q. and Zhou, W., 2017. Testing for high-dimensional white noise using maximum cross-correlations. Biometrika, 104(1), pp.111-127.

Cai, T.T., Liu, W., and Luo, X. (2011). A constrained `1 minimization approach for sparse precision matrix estimation. Journal of the American Statistical Association 106(494): 594-607.

L\"utkepohl, Helmut. New introduction to multiple time series analysis. Springer Science & Business Media, 2005.
}
\author{
Meng Cao, Wen Zhou
}
